Final wrapper tests

+++ SensorUtilities/include/MathUtils.h
@@ -27,7 +27,7 @@ namespace SensorUtilities {
   struct ImagePt {
     double line;
     double sample;
-    double band;
+    int band;
   };
